# ecommerce-site-laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Commerce API",
        "description": "Documentation for E-Commerce API with Delivery Management",
        "contact": {
            "email": "admin@ecommerce.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        },
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/admin/dashboard/overview": {
            "get": {
                "tags": [
                    "Admin Dashboard"
                ],
                "summary": "Get dashboard overview stats",
                "description": "Obtient les statistiques globales du dashboard\n\nGET /api/admin/dashboard/overview",
                "operationId": "getDashboardOverview",
                "parameters": [
                    {
                        "name": "period",
                        "in": "query",
                        "description": "Period (today, week, month, year, all)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "today"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dashboard statistics",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "period": {
                                                    "type": "string",
                                                    "example": "today"
                                                },
                                                "statistics": {
                                                    "type": "object"
                                                },
                                                "generated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/dashboard/sales-chart": {
            "get": {
                "tags": [
                    "Admin Dashboard"
                ],
                "summary": "Get sales chart data",
                "description": "Obtient les données pour le graphique d'évolution des ventes\n\nGET /api/admin/dashboard/sales-chart",
                "operationId": "getSalesChart",
                "parameters": [
                    {
                        "name": "period",
                        "in": "query",
                        "description": "Period (week, month, year)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "month"
                        }
                    },
                    {
                        "name": "group_by",
                        "in": "query",
                        "description": "Group by (day, week, month)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "day"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sales chart data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "labels": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                },
                                                "orders": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "integer"
                                                    }
                                                },
                                                "revenue": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "number"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/dashboard/top-products": {
            "get": {
                "tags": [
                    "Admin Dashboard"
                ],
                "summary": "Get top selling products",
                "description": "Obtient la liste des produits les plus vendus\n\nGET /api/admin/dashboard/top-products",
                "operationId": "getTopProducts",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of products",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "period",
                        "in": "query",
                        "description": "Period (today, week, month, year)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "month"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Top products list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "image_url": {
                                                        "type": "string"
                                                    },
                                                    "total_sold": {
                                                        "type": "integer"
                                                    },
                                                    "total_revenue": {
                                                        "type": "string"
                                                    },
                                                    "total_revenue_raw": {
                                                        "type": "number"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/admin/dashboard/recent-activity": {
            "get": {
                "tags": [
                    "Admin Dashboard"
                ],
                "summary": "Get recent activity feed",
                "description": "Obtient l'activité récente du système\n\nGET /api/admin/dashboard/recent-activity",
                "operationId": "getRecentActivity",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit results",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recent activity list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "type": {
                                                        "type": "string",
                                                        "example": "new_order"
                                                    },
                                                    "title": {
                                                        "type": "string"
                                                    },
                                                    "message": {
                                                        "type": "string"
                                                    },
                                                    "timestamp": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "time_ago": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}